initial upload
This version doesn't pan the audio very well - I'm still learning that
part. The main reason for uploading at this point in development is to
double-check the CORS and other potential problems with moving to using
a global file base for the general files, such as input handling,
storage, etc.